Migrate account settings page
